backend/requirements.txt
flask
flask-sqlalchemy
flask-cors
prophet
pandas
python-dotenv
